+++ software/CanLinConfig/Helpers/BindingProxy.cs
@@ -1,0 +1,17 @@
+using System.Windows;
+
+namespace CanLinConfig.Helpers;
+
+public class BindingProxy : Freezable
+{
+    public static readonly DependencyProperty DataProperty =
+        DependencyProperty.Register(nameof(Data), typeof(object), typeof(BindingProxy));
+
+    public object Data
+    {
+        get => GetValue(DataProperty);
+        set => SetValue(DataProperty, value);
+    }
+
+    protected override Freezable CreateInstanceCore() => new BindingProxy();
+}